requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
faiss-cpu>=1.8.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
numpy>=1.26.0,<3.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.35.0,<2.0.0
google-genai>=1.11.0,<2.0.0
rank-bm25>=0.2.2,<1.0.0
PyYAML>=6.0.1,<7.0.0
tqdm>=4.66.0,<5.0.0
